Improve script

+++ build.sh
@@ -1,39 +1,50 @@
 #!/bin/bash
 
-TOOLCHAIN_DIR=/home/skia/workspace/toolchains
-TOOLCHAIN_BUILD_DIR=${TOOLCHAIN_DIR}/build/
-TOOLCHAIN_BR_DIR=${TOOLCHAIN_BUILD_DIR}/buildroot
+TOOLCHAIN_DIR=~/toolchains-builder/
+TOOLCHAIN_BUILD_DIR=${TOOLCHAIN_DIR}/builds
+TOOLCHAIN_BR_DIR=${TOOLCHAIN_DIR}/buildroot
 TOOLCHAIN_VERSION=$(git --git-dir=${TOOLCHAIN_BR_DIR}/.git describe)
 
 function set_qemu_config {
     if [ ${arch} == "arm" ]; then
         qemu_defconfig="qemu_arm_versatile_defconfig"
-        qemu_machine="versatilepb"
-        qemu_dtb="versatile-pb.dtb"
+        qemu_system_command="qemu-system-arm
+            -machine versatilepb
+            -kernel ${testdir}/images/zImage
+            -dtb ${testdir}/images/versatile-pb.dtb
+            -drive file=${testdir}/images/rootfs.ext2,index=0,media=disk
+            -append \"root=/dev/sda rw\""
     elif [ ${arch} == "aarch64" ]; then
         qemu_defconfig="qemu_aarch64_virt_defconfig"
-        qemu_machine="virt"
-        qemu_dtb="virt.dtb"
+        qemu_system_command="qemu-system-aarch64
+            -machine virt -cpu cortex-a57 -machine type=virt
+            -kernel ${testdir}/images/Image
+            -initrd ${testdir}/images/rootfs.cpio
+            -append \"root=/dev/sda rw\""
+    elif [ ${arch} == "mipsel" ]; then
+        qemu_defconfig="qemu_mips32r2_malta_defconfig"
+        qemu_system_command="qemu-system-mipsel
+            -machine malta
+            -kernel ${testdir}/images/Image
+            -drive file=${testdir}/images/rootfs.cpio,index=0,media=disk
+            -append \"root=/dev/sda rw\""
     fi
-}
-function eval_command() {
-  "$@";
 }
 
 function boot_test {
-    # cd ${testdir}/images
-    macaddress="f8:ca:b8:3f:ae:e9"
+    # cd ${testdir}/images
     echo "  booting test system ... "
-    qemu-system-${arch} \
-        -machine ${qemu_machine} \
-        -kernel ${testdir}/images/zImage \
-        -dtb ${testdir}/images/${qemu_dtb} \
-        -hda ${testdir}/images/rootfs.ext2 \
-        -append "root=/dev/sda rw" \
-        -serial telnet:127.0.0.1:4000,server,nowait,nodelay \
-        -nographic &
+    qemu_system_command="${qemu_system_command}
+            -serial telnet:127.0.0.1:4000,server,nowait,nodelay
+            -nographic"
+    echo "  boot command: ${qemu_system_command}"
+    eval ${qemu_system_command} &
     echo $! > /tmp/qemu-test.pid
     sleep 2
+    if ! ps p $(cat /tmp/qemu-test.pid) &>/dev/null; then
+        echo "  Failed to launch qemu ... Not going further"
+        return 1
+    fi
     if ! expect expect.sh; then
         echo "  booting test system ... FAILED"
         kill -9 $(cat /tmp/qemu-test.pid)
@@ -90,7 +101,7 @@ function build {
     mkdir ${builddir}
 
     # Create the configuration
-    cp ${TOOLCHAIN_BUILD_DIR}/${name}.config ${configfile}
+    cp ${TOOLCHAIN_DIR}/${name}.config ${configfile}
     echo "BR2_JLEVEL=16" >> ${configfile}
     echo "BR2_HOST_DIR=\"${toolchaindir}\"" >> ${configfile}
 
@@ -115,11 +126,11 @@ function build {
 
 function generate {
     name=$1
-    toolchaindir=/opt/${name}-${TOOLCHAIN_VERSION}
+    toolchaindir=${TOOLCHAIN_BUILD_DIR}/${name}-${TOOLCHAIN_VERSION}
     testdir=${toolchaindir}-tests
-    logfile=/opt/${name}-${TOOLCHAIN_VERSION}-build.log
-    testlogfile=/opt/${name}-${TOOLCHAIN_VERSION}-test.log
-    builddir=/opt/toolchain-build/
+    logfile=${TOOLCHAIN_BUILD_DIR}/${name}-${TOOLCHAIN_VERSION}-build.log
+    testlogfile=${TOOLCHAIN_BUILD_DIR}/${name}-${TOOLCHAIN_VERSION}-test.log
+    builddir=${TOOLCHAIN_BUILD_DIR}/toolchain-build/
     configfile=${builddir}/.config
 
     echo "Generating ${name}..."
@@ -141,11 +152,11 @@ function generate {
     boot_test
     return
 
-    # Everything works, package the toolchain
-    (cd /opt; tar cjf `basename ${toolchaindir}`.tar.bz2 `basename ${toolchaindir}`)
+    # Everything works, package the toolchain
+    (cd ${TOOLCHAIN_BUILD_DIR}; tar cjf `basename ${toolchaindir}`.tar.bz2 `basename ${toolchaindir}`)
 
     # Remove toolchain directory
-    # rm -rf ${toolchaindir}
+    # rm -rf ${toolchaindir}
 
     # Remove build directory
     # rm -rf ${builddir}
